requirements.txt
basicsr==1.4.2
config==0.5.1
dataset==1.6.2
deprecated==1.2.14
fire==0.5.0
fpdf==1.7.2
gfpgan==1.3.8
loss==0.1.2
mailjet_rest==1.3.4
model==0.6.0
mysql_connector_repackaged==0.3.1
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
Pillow==10.1.0
realesrgan==0.3.0
streamlit==1.21.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
utils==1.0.1
